+++ testapp/testapp/testmain/fixtures.py
@@ -109,7 +109,6 @@ class ReceiptFactory(DjangoModelFactory):
     concept = SubFactory(ConceptTypeFactory)
     document_type = SubFactory(DocumentTypeFactory)
     document_number = 33445566
-    expiration_date = datetime(2017, 7, 25)
     issued_date = datetime(2017, 5, 15)
     total_amount = 100
     net_untaxed = 0